requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.5
django-cart @ git+https://github.com/bmentges/django-cart.git@f2578e7cbe00a64bde3139c6da7c05f820cd281a
django-summernote==0.8.11.6
gunicorn==20.1.0
idna==2.10
Pillow==8.3.2
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.2
stripe==2.57.0
urllib3==1.26.5
whitenoise==5.2.0
